benchmark refactoring
